+++ che-theia-factory/src/browser/types.ts
